# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Second Project API",
    "description": "CSE 341 Week 04 API with CRUD, validation, error handling, and GitHub OAuth authentication",
    "version": "1.0.0"
  },
  "host": "cse341-second-projects.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "cookieAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "connect.sid",
      "description": "Session cookie returned after successful OAuth login."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/oauth/github/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "GitHub OAuth callback endpoint.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/books/": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "The Pragmatic Programmer"
                },
                "author": {
                  "type": "string",
                  "example": "Andrew Hunt"
                },
                "genre": {
                  "type": "string",
                  "example": "Software Development"
                },
                "publishedYear": {
                  "type": "number",
                  "example": 1999
                },
                "pages": {
                  "type": "number",
                  "example": 352
                },
                "rating": {
                  "type": "number",
                  "example": 4.8
                },
                "inStock": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/receipts/": {
      "get": {
        "tags": [
          "Receipts"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Receipts"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "receiptNumber": {
                  "type": "string",
                  "example": "REC-2026-001"
                },
                "vendorName": {
                  "type": "string",
                  "example": "The Book Haven"
                },
                "receiptDate": {
                  "type": "string",
                  "example": "2026-03-28"
                },
                "totalAmount": {
                  "type": "number",
                  "example": 125.5
                },
                "itemCount": {
                  "type": "number",
                  "example": 5
                },
                "paymentMethod": {
                  "type": "string",
                  "example": "credit_card"
                },
                "category": {
                  "type": "string",
                  "example": "groceries"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  }
}